refactor: move JWSGenerator up
